Adding new template "post-template-full"
This new template allows a difference in the layout of posts on the homepage post list and posts being read on their own pages. If you do not include a new <script id="post-template-full"> template in your index.html, CMS.js will fall back to using "#post-template" for both instances.